🔧 Configure PyUp bot to run monthly; prefix emoji (#35)
The PyUp bot currently runs with defaults: as soon as a new package is
updated, the bot opens a new PR. Given the size and stability of this
package, that's overkill and takes up more of my time (multiple PRs when
we could use one).

This update limits the bot to a monthly schedule. We also configure the
bot to prefix pull-requests with an emoji. For reasons.