Enable short name via entrypoint (#45)
Markdown's behavior pre-v3 was to automatically append mdx_ to strings
that acted as references to third-party extensions. Starting in Markdown
v3, this behavior has been removed. In v2.1.0 of this package, we
therefore lost the ability to refer to superscript (forcing developers
to switch to mdx_superscript) if referring to the extension by string
(but only in Markdown v3).

As it turns out, we can re-enable the old behavior by using an
entrypoint in setup.py. This PR therefore re-enables previous behavior.

* ❗️ Upgrade tox 3.4.0 to 3.5.2
* 🚀 Improve Makefile clean target
* ✨ Add entrypoint; 🚀 improve setup.py

Setup.py now:

1. defines a Markdown entrypoint, allowing for string reference
   by short module (see docs)
2. adds extra links for use on PyPI
3. displays only relevant section of ReadMe on PyPI
4. cleans up code (constants are upper-cased, duplicate code refactored)